% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eindim618funk.R
\name{absolutabweichungfunk}
\alias{absolutabweichungfunk}
\title{caculate the absolut abweichung}
\usage{
absolutabweichungfunk(A, B)
}
\arguments{
\item{A}{vector, numica}

\item{B}{vector, numica}
}
\value{
Num, sum(abs(A-B))
}
\description{
caculate the absolut abweichung
}
\examples{
absolutabweichungfunk(c(1,5), c(2,3))
}
