% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eindim618funk.R
\name{absolutabweichungVector}
\alias{absolutabweichungVector}
\title{caculate the absolut abweichung for vector}
\usage{
absolutabweichungVector(A, B)
}
\arguments{
\item{A}{vector, numica}

\item{B}{vector, numica}
}
\value{
Num, abs(A-B)
}
\description{
caculate the absolut abweichung for vector
}
\examples{
absolutabweichungfunk(c(1,5), c(2,3))
}
