% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eindim618funk.R
\name{eindim618funk}
\alias{eindim618funk}
\title{optimisation with 0.618}
\usage{
eindim618funk(
  processfk,
  paramekl,
  paramegr,
  DatenReferenz,
  abweichfk = absolutabweichungfunk,
  umkriesn,
  ...
)
}
\arguments{
\item{processfk}{Function, Self function}

\item{paramekl}{numica, the small wert}

\item{paramegr}{numica, the big wert}

\item{DatenReferenz}{numica, the Ideal optimal solution}

\item{abweichfk}{Function, Deviation calculation function, default absolute error}

\item{umkriesn}{Num, how many search}

\item{...}{other paramater for processfk}
}
\value{
Num, the best wert Optimal solution
}
\description{
optimisation with 0.618
}
