% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eindim618funk.R
\name{eindim618Vector}
\alias{eindim618Vector}
\title{optimisation with 0.618 for vector}
\usage{
eindim618Vector(
  processfk,
  LowerBound,
  UpperBound,
  idealReferenz,
  abweichfk = absolutabweichungVector,
  umkriesn = 5,
  ...
)
}
\arguments{
\item{processfk}{Function, Self function}

\item{LowerBound}{numica vector, the small wert}

\item{UpperBound}{numica vector, the big wert}

\item{idealReferenz}{numica or vector, the Ideal optimal solution}

\item{abweichfk}{Function, Deviation calculation function, default absolute error}

\item{umkriesn}{Num, how many search}

\item{...}{other paramater for processfk}
}
\value{
numica vector, the best wert Optimal solution
}
\description{
optimisation with 0.618 for vector
}
